Output to STDOUT and STDERR directly when running specs
Rather than wait for all specs to run and complete before outputting the
spec results, have the process running pipe the results directly to
STDOUT and STDERR when running the specs specifically. Does not do this
if running in silent mode.

This way if the specs are running on a timeout that expires and kills
the crkcov process running the specs, we don't lose all output / error
messages that may have already occurred (but were being buffered)